src/main/java/io/homo_efficio/learnmicroservicesspringboot/multiplication/service/MultiplicationServiceImpl.java
package io.homo_efficio.learnmicroservicesspringboot.multiplication.service;

import io.homo_efficio.learnmicroservicesspringboot.multiplication.domain.Multiplication;
import io.homo_efficio.learnmicroservicesspringboot.multiplication.domain.MultiplicationAttempt;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

/**
 * @author homo.efficio@gmail.com
 * created on 2019-10-06
 */
@Service
@RequiredArgsConstructor
public class MultiplicationServiceImpl implements MultiplicationService {

    private final RandomGeneratorService randomGeneratorService;

    @Override
    public Multiplication createRandomMultiplication() {
        int factorA = randomGeneratorService.generateRandomFactor();
        int factorB = randomGeneratorService.generateRandomFactor();
        return new Multiplication(factorA, factorB);
    }

    @Override
    public boolean checkAttempt(MultiplicationAttempt resultAttempt) {
        return false;
    }
}
